+++ packages/react-app-renderer/src/server.tsx
@@ -2,9 +2,11 @@ import * as React from 'react';
 import * as ReactDOMServer from 'react-dom/server';
 import { ChunkExtractor } from '@loadable/server';
 import { getRenderApp } from './renderer';
+import type { Context, RenderOptions } from './types';
 
-function renderInServer(context, options) {
-  const { appConfig, buildConfig = {}, createBaseApp, emitLifeCycles } = options;
+function renderInServer(context: Context, options: RenderOptions) {
+  const { appConfig, buildConfig = {}, appLifecycle } = options;
+  const { createBaseApp, emitLifeCycles } = appLifecycle;
   const { runtime, appConfig: modifiedAppConfig } = createBaseApp(appConfig, buildConfig, context);
 
   const { loadableStatsPath, publicPath } = buildConfig;
@@ -28,7 +30,7 @@ function renderInServer(context, options) {
   };
 }
 
-export default function reactAppRendererWithSSR(context, options) {
+export default function reactAppRendererWithSSR(context: Context, options: RenderOptions) {
   const cloneOptions = deepClone(options);
   const { appConfig } = cloneOptions || {};
   appConfig.router = appConfig.router || {};
